+++ qtah-cpp/Setup.hs
@@ -91,7 +91,10 @@ import Distribution.Simple.Utils (
   installOrdinaryFile,
   notice,
   )
-#if MIN_VERSION_Cabal(2,2,0)
+#if MIN_VERSION_Cabal(3,2,0)
+-- GHC 8.10.1+:
+import Distribution.Types.Flag (lookupFlagAssignment)
+#elif MIN_VERSION_Cabal(2,2,0)
 import Distribution.Types.GenericPackageDescription (lookupFlagAssignment)
 #endif
 import Distribution.Verbosity (Verbosity, normal)