+++ qtah/Setup.hs
@@ -97,7 +97,10 @@ import Distribution.Simple.Utils (die')
 import Distribution.Simple.Utils (die)
 #endif
 import Distribution.Simple.Utils (info, installOrdinaryFile, notice)
-#if MIN_VERSION_Cabal(2,2,0)
+#if MIN_VERSION_Cabal(3,2,0)
+-- GHC 8.10.1+:
+import Distribution.Types.Flag (lookupFlagAssignment)
+#elif MIN_VERSION_Cabal(2,2,0)
 import Distribution.Types.GenericPackageDescription (lookupFlagAssignment)
 #endif
 import Distribution.Verbosity (Verbosity, normal)